# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "9router API",
    "version": "0.4.71",
    "description": "Public LLM API for 9router (`/v1/*`), compatible with OpenAI- and Claude-style clients. Authenticate with a 9router API key via `Authorization: Bearer <key>` or `x-api-key: <key>`."
  },
  "servers": [
    {
      "url": "http://localhost:20128",
      "description": "Local dev"
    },
    {
      "url": "http://35.255.190.170:20128",
      "description": "Develop"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "9router API key sent as `Authorization: Bearer <key>` (OpenAI-style clients)."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "9router API key sent as `x-api-key: <key>` (Anthropic/Claude-style clients)."
      },
      "CliTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-9r-cli-token",
        "description": "Local CLI token (machineId-derived). Accepted on /v1 and /api/* from the local CLI."
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth_token",
        "description": "Dashboard session JWT set after login. Used by the management /api/* endpoints."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "oneOf": [
              {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "example": "invalid_request_error"
                  }
                },
                "required": [
                  "message"
                ]
              },
              {
                "type": "string"
              }
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Model": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "openai/gpt-4o"
          },
          "object": {
            "type": "string",
            "example": "model"
          },
          "owned_by": {
            "type": "string",
            "description": "Provider alias or `combo`."
          },
          "kind": {
            "type": "string",
            "description": "Only present for non-LLM models (e.g. webSearch, webFetch)."
          }
        }
      },
      "ModelList": {
        "type": "object",
        "properties": {
          "object": {
            "type": "string",
            "example": "list"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Model"
            }
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/v1/models/{kind}": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List models by capability",
        "description": "OpenAI-compatible model list filtered by service kind.",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          },
          {
            "CliTokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "image",
                "tts",
                "stt",
                "embedding",
                "image-to-text",
                "web"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of models for the requested kind.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelList"
                }
              }
            }
          },
          "404": {
            "description": "Unknown model kind.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/models": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List available LLM models",
        "description": "OpenAI-compatible model list (LLM/chat models only). For other capabilities use /v1/models/{kind}.",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          },
          {
            "CliTokenAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of models.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelList"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}